+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - P2P can now work without adding the p2p4slips binary to PATH
 - Add detection for connections to private IPs from private IPs
 - Add detection for DHCP scans
+- Support having IP ranges in your own local TI file own_malicious_iocs.csv
 -1.0.0
 - ignore NXDOMAINs dns resolution when checking for 'dns without resolutions'
 - Keep track of old peer reports about the same ip